# config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "v1",
        "title": "Contact List",
        "description": "A Contact list API based on Swagger and built using Node.js"
    },
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "operationId": "users_get",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                },
                "deprecated": false
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "operationId": "users_add",
                "consumes": [
                    "application/json",
                    "text/json"
                ],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserAdd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "deprecated": false
            }
        },
        "/users/username/{username}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "operationId": "users_getByUsername",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/users/userid/{user_id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "operationId": "users_getById",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "operationId": "events_getAll",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Event"
                            }
                        }
                    }
                },
                "deprecated": false
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "operationId": "events_add",
                "consumes": [
                    "application/json",
                    "text/json"
                ],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/EventAdd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "deprecated": false
            }
        },
        "/events/getById/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "operationId": "events_getById",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Event"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/events/getByKeyword/{keyword}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "operationId": "events_getByKeyword",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Event"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/events/attendees/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "operationId": "events_getAttendees",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Event"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/events/getByUser/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "operationId": "events_getByUser",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Event"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/events/getByRadius": {
            "get": {
                "tags": [
                    "Events"
                ],
                "operationId": "events_getByRadius",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Event"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/events/getInitial": {
            "get": {
                "tags": [
                    "Events"
                ],
                "operationId": "events_getInitial",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Event"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/events/rsvp": {
            "get": {
                "tags": [
                    "Events"
                ],
                "operationId": "events_rsvp",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "userid",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "eventid",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Event"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/events/rsvpStatus": {
            "get": {
                "tags": [
                    "Events"
                ],
                "operationId": "events_rsvpStatus",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "userid",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "eventid",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Event"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/contacts": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "operationId": "contacts_get",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Contact"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/contacts/{id}": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "operationId": "contacts_getById",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Contact"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        }
    },
    "definitions": {
        "User": {
            "type": "object",
            "properties": {
                "user_id": {
                    "format": "int32",
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                },
                "screen_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "university": {
                    "type": "string"
                },
                "pword": {
                    "type": "string"
                }
            }
        },
        "UserAdd": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                },
                "screen_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "university": {
                    "type": "string"
                },
                "pword": {
                    "type": "string"
                }
            }
        },
        "Event": {
            "type": "object",
            "properties": {
                "event_id": {
                    "format": "int32",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "num_attending": {
                    "format": "int32",
                    "type": "integer"
                },
                "university": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "lat": {
                    "type": "string"
                },
                "lng": {
                    "type": "string"
                }
            }
        },
        "EventAdd": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "university": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "lat": {
                    "type": "string"
                },
                "lng": {
                    "type": "string"
                }
            }
        },
        "Contact": {
            "type": "object",
            "properties": {
                "id": {
                    "format": "int32",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        }
    }
}